services/notification-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.30.0
structlog==24.4.0
python-jose[cryptography]==3.3.0
resend==2.4.0
aiohttp==3.10.10
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
python-dotenv==1.0.1
jinja2==3.1.4
